refactor(datepicker): fix, remove events, ui, ...
Description:
- moved input inside component to simplify and contain logic
- added store for (from/to dates) so that the catalogue can retrieve the dates
- fix ui : the datepicker now appears on top, changed color for month and year
- refactor : removed events, added custom directive
- changed behaviour : when the modal is not open it is not in the DOM
- moved translations elsewhere to keep the .vue file "concise"
- implemented simple test for future improvements
