+++ src/lib/modules/aeris-datepicker/assets/dates.json
@@ -35,7 +35,7 @@
     "october": "Octobre",
     "november": "Novembre",
     "december": "Décembre",
-    "today": "Aujourdhui",
+    "today": "Aujourd'hui",
     "validate": "Valider",
     "mon": "Lun",
     "tue": "Mar",
